Test on the influence of capacitance on +3.3V voltage supply.
It is obvious that the power supply has zigzag noise after the capacitor is removed. After adding some small capacitors, the noise almostdisappears. After adding all the capacitors, the noise is difficult to display on the oscilloscope. The random two sampling points have a voltage of 3.28V, which is close to desired 3.3V.